Write duplicated modules using macros
